requirements.txt
plotly==5.9.0
pandas==2.0.3
numpy==1.24.3
folium==0.15.1
haversine==2.8.0
streamlit==1.29.0
Pillow==9.4.0
IPython==8.12.0
streamlit-folium==0.15.1
geopy==2.4.1